Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)

The reason I was planning to add usage for these callbacks in a later commit (or amending this commit when we are done with updates which have a higher priority) is that we'd need to do some refactoring regarding best practices throughout the code. Considering that we might add new functions during this refactoring process, I was more inclined to not add these callbacks for now. However, my assessment might be wrong - waiting on answer.

Patch-set: 2
Reviewer: Gerrit User 1016453 <1016453@d5d70762-12d0-45a1-890d-524b12d3f735>