Fix getData and setData
Update messageHandler, parseULMessage, and parseULValue (renamed
to parseStringValue:

 - messageHandler: now calls the listDevices() method in order to
pass devices as a parameter for parseULMessage (which doesn't have
access to the method).

 - parseULMessage: 'devices' was added as parameter so that the
sensor's valueType is accessible and passed on as parameter for
parseULValue.

 - parseULValue: as comlpex commands (w/ multiple parameters) are not
supported, code which assesed this was removed. Furthermore, since
knot-base-connector does a type validation on incoming messages, logic
was added to convert 'value' (string) to it's sensor's valueType.

Change-Id: I579a3f53d5457ab97c8cc925392b1242e7926cbe