Add updateSchema error handling
The 409 error means the request couldn't be completed due to a conflict in
respect to the target resource - contextually, a (deviceId, serviceId) pair
already exists. However, this occurs when attempting to make changes in a
resource which the base connector is already acting upon due to a previous
request (e.g. updateSchema). As soon as the request is processed, the error
will be gone, without consequences; therefore, it shouldn't case an interuption.

Change-Id: I046e576e24d635766be9edc46c3c56e93bdb2656