Remove Error 409 handling from updateSchema
The try-catch block in updateSchema() aims to handle Error 409
(duplicate device). However, the error is caused by an error in
deviceExistsOnIoTA() and was corrected in commit 549bd25.

As a result, the comment (describing what was believed to be the
cause of the error) is incorrect and the  try-catch block handles
nothing -- just delete them.